Move LoadoutsRow to its own file
