+++ src/app/loadout/Loadouts.m.scss
@@ -24,7 +24,3 @@
   composes: dim-button from global;
   text-align: left;
 }
-
-.iconLabel {
-  margin-right: 8px;
-}
