+++ src/app/loadout/Loadouts.tsx
@@ -3,18 +3,14 @@ import { DestinyAccount } from 'app/accounts/destiny-account';
 import { apiPermissionGrantedSelector, languageSelector } from 'app/dim-api/selectors';
 import { AlertIcon } from 'app/dim-ui/AlertIcon';
 import CharacterSelect from 'app/dim-ui/CharacterSelect';
-import { ConfirmButton } from 'app/dim-ui/ConfirmButton';
 import FilterPills, { Option } from 'app/dim-ui/FilterPills';
 import PageWithMenu from 'app/dim-ui/PageWithMenu';
 import ShowPageLoading from 'app/dim-ui/ShowPageLoading';
 import { t, tl } from 'app/i18next-t';
 import { getHashtagsFromNote } from 'app/inventory/note-hashtags';
 import { sortedStoresSelector } from 'app/inventory/selectors';
-import { DimStore } from 'app/inventory/store-types';
 import { useLoadStores } from 'app/inventory/store/hooks';
 import { getCurrentStore, getStore } from 'app/inventory/stores-helpers';
-import { deleteLoadout } from 'app/loadout-drawer/actions';
-import { applyLoadout } from 'app/loadout-drawer/loadout-apply';
 import { editLoadout } from 'app/loadout-drawer/loadout-events';
 import { Loadout } from 'app/loadout-drawer/loadout-types';
 import {
@@ -25,29 +21,19 @@ import {
 import { loadoutsSelector } from 'app/loadout-drawer/selectors';
 import { plainString } from 'app/search/search-filters/freeform';
 import { useSetting } from 'app/settings/hooks';
-import {
-  addIcon,
-  AppIcon,
-  deleteIcon,
-  faCalculator,
-  faCheckCircle,
-  uploadIcon,
-} from 'app/shell/icons';
+import { addIcon, AppIcon, faCalculator, uploadIcon } from 'app/shell/icons';
 import { querySelector, useIsPhonePortrait } from 'app/shell/selectors';
-import { useThunkDispatch } from 'app/store/thunk-dispatch';
-import { streamDeckSelectionSelector } from 'app/stream-deck/selectors';
-import { streamDeckSelectLoadout } from 'app/stream-deck/stream-deck';
 import { Portal } from 'app/utils/temp-container';
 import { DestinyClass } from 'bungie-api-ts/destiny2';
 import deprecatedMods from 'data/d2/deprecated-mods.json';
 import _ from 'lodash';
-import { ReactNode, useMemo, useState } from 'react';
+import { useMemo, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { Link } from 'react-router-dom';
 import LoadoutImportSheet from './loadout-share/LoadoutImportSheet';
 import LoadoutShareSheet from './loadout-share/LoadoutShareSheet';
 import styles from './Loadouts.m.scss';
-import LoadoutView from './LoadoutView';
+import LoadoutRow from './LoadoutsRow';
 
 const sortOptions = [
   {
@@ -278,93 +264,5 @@ function Loadouts({ account }: { account: DestinyAccount }) {
         </Portal>
       )}
     </PageWithMenu>
-  );
-}
-
-// TODO: memoize?
-function LoadoutRow({
-  loadout,
-  store,
-  saved,
-  equippable,
-  onShare,
-}: {
-  loadout: Loadout;
-  store: DimStore;
-  saved: boolean;
-  equippable: boolean;
-  onShare: (loadout: Loadout) => void;
-}) {
-  const dispatch = useThunkDispatch();
-
-  const streamDeckSelection = $featureFlags.elgatoStreamDeck
-    ? // eslint-disable-next-line
-      useSelector(streamDeckSelectionSelector)
-    : null;
-
-  const actionButtons = useMemo(() => {
-    const handleDeleteClick = (loadout: Loadout) => dispatch(deleteLoadout(loadout.id));
-
-    const handleApply = () =>
-      dispatch(applyLoadout(store, loadout, { allowUndo: true, onlyMatchingClass: true }));
-
-    const handleEdit = () => editLoadout(loadout, store.id, { isNew: !saved });
-    const actionButtons: ReactNode[] = [];
-
-    if (equippable) {
-      if (streamDeckSelection === 'loadout') {
-        const handleSelection = () => dispatch(streamDeckSelectLoadout(loadout, store));
-        return [
-          <button
-            key="select-for-stream-deck"
-            type="button"
-            className="dim-button"
-            onClick={handleSelection}
-          >
-            <span className={styles.iconLabel}>{t('StreamDeck.SelectLoadout')}</span>
-            <AppIcon icon={faCheckCircle} title={t('StreamDeck.SelectLoadout')} />
-          </button>,
-        ];
-      }
-
-      actionButtons.push(
-        <button key="apply" type="button" className="dim-button" onClick={handleApply}>
-          {t('Loadouts.Apply')}
-        </button>
-      );
-    }
-
-    actionButtons.push(
-      <button key="edit" type="button" className="dim-button" onClick={handleEdit}>
-        {saved ? t('Loadouts.EditBrief') : t('Loadouts.SaveLoadout')}
-      </button>
-    );
-
-    if (loadout.parameters && !_.isEmpty(loadout.parameters)) {
-      actionButtons.push(
-        <button key="share" type="button" className="dim-button" onClick={() => onShare(loadout)}>
-          {t('Loadouts.ShareLoadout')}
-        </button>
-      );
-    }
-
-    if (saved) {
-      actionButtons.push(
-        <ConfirmButton key="delete" danger onClick={() => handleDeleteClick(loadout)}>
-          <AppIcon icon={deleteIcon} title={t('Loadouts.Delete')} />
-        </ConfirmButton>
-      );
-    }
-
-    return actionButtons;
-  }, [dispatch, equippable, loadout, onShare, saved, store, streamDeckSelection]);
-
-  return (
-    <LoadoutView
-      loadout={loadout}
-      store={store}
-      actionButtons={actionButtons}
-      hideShowModPlacements={!equippable}
-    />
   );
 }
