+++ src/app/loadout/LoadoutsRow.m.scss
@@ -1,0 +1,3 @@
+.iconLabel {
+  margin-right: 8px;
+}
