+++ src/app/loadout/LoadoutsRow.m.scss.d.ts
@@ -1,10 +1,7 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
-  'menu': string;
-  'menuButton': string;
-  'menuButtons': string;
-  'page': string;
+  'iconLabel': string;
 }
 export const cssExports: CssExports;
 export default cssExports;
